fix: remove duplicate shebang
