+++ scripts/hashtags_to_kanban.py
@@ -1,73 +1,7 @@
 #!/usr/bin/env python3
 """Generate a Kanban board from task status hashtags."""
 
-from __future__ import annotations
-
-import re
-from collections import defaultdict
-from pathlib import Path
-
-TASK_DIR = Path("docs/agile/tasks")
-
-# Ordered list of recognized status hashtags
-STATUS_ORDER = [
-    "#ice-box",
-    "#incoming",
-    "#rejected",
-    "#accepted",
-    "#prompt-refinement",
-    "#agent-thinking",
-    "#breakdown",
-    "#blocked",
-    "#ready",
-    "#todo",
-    "#in-progress",
-    "#in-review",
-    "#done",
-]
-STATUS_SET = set(STATUS_ORDER)
-
-TITLE_RE = re.compile(r"^##\s+🛠️\s+Task:\s*(.+)")
-HASHTAG_RE = re.compile(r"#([A-Za-z0-9_-]+)")
-
-
-def parse_task(path: Path) -> tuple[str, str]:
-    """Return task title and status hashtag from a markdown file."""
-    title = path.stem.replace("_", " ")
-    status = "#todo"
-    with path.open(encoding="utf-8") as fh:
-        for line in fh:
-            if m := TITLE_RE.match(line):
-                title = m.group(1).strip()
-            for tag in HASHTAG_RE.findall(line):
-                tag = f"#{tag}"
-                if tag in STATUS_SET:
-                    status = tag
-    return title, status
-
-
-def collect_tasks(directory: Path = TASK_DIR):
-    tasks: dict[str, list[tuple[str, Path]]] = defaultdict(list)
-    for file in directory.glob("*.md"):
-        title, status = parse_task(file)
-        tasks[status].append((title, file))
-    return tasks
-
-
-def build_board(tasks: dict[str, list[tuple[str, Path]]]) -> str:
-    lines = ["---", "", "kanban-plugin: board", "", "---", ""]
-    for status in STATUS_ORDER:
-        items = tasks.get(status)
-        if not items:
-            continue
-        header = status.lstrip("#").replace("-", " ").title()
-        lines.append(f"## {header}")
-        lines.append("")
-        for title, path in sorted(items):
-            rel = Path("../tasks") / path.name
-            lines.append(f"- [ ] [{title}]({rel}) {status}")
-        lines.append("")
-    return "\n".join(lines)
+from shared.py.agile.kanban import build_board, collect_tasks
 
 
 def main() -> None:
