+++ shared/py/agile/__init__.py
@@ -1,0 +1,25 @@
+"""Utilities for working with agile task boards."""
+
+from .kanban import (
+    BOARD_PATH,
+    TASK_DIR,
+    STATUS_ORDER,
+    STATUS_SET,
+    build_board,
+    collect_tasks,
+    parse_board,
+    set_status,
+    update_tasks,
+)
+
+__all__ = [
+    "BOARD_PATH",
+    "TASK_DIR",
+    "STATUS_ORDER",
+    "STATUS_SET",
+    "build_board",
+    "collect_tasks",
+    "parse_board",
+    "set_status",
+    "update_tasks",
+]
